Add initial support for custom parameter injection
Major refactoring to support change and improve clarity

Currently to delirious to write a proper comment... maybe I will rewrite this later

Custom parameter injection can be useful when the you want version of the same type configured in a different way.  A common example of this might be Date or Number formatters.