Remove need for required constructors Remove the need for multiple protocol variants to support value and reference types Lifetime is now determined based on conformance to LifetimeProviding protocol rather than specification of lifetime parameter VariantCreating protocol allows for creation of variants of the same type differentiated by a string key Maintained cycle support
This implementation relies heavily on generics which seem to still have some issues based on segmentation faults observed on compilation rather than the compiler producing errors.  This seems to happen if you do not correctly conform to the protocol in some specific ways I have not bothered to isolate yet.  Functions correctly when specified correctly but could be annoying for users if the reason is not clear.